move operation tracking workflow code to operation module
