+++ src/maastemporalworker/workflow/utils.py
@@ -2,7 +2,6 @@
 # GNU Affero General Public License version 3 (see the file LICENSE).
 
 import asyncio
-from datetime import timedelta
 from functools import wraps
 import random
 
@@ -10,13 +9,10 @@ import structlog
 from temporalio import activity, workflow
 from temporalio.exceptions import ApplicationError, TemporalError
 
-from maascommon.enums.operations import OperationStatus
 from maascommon.tracing import get_or_set_trace_id
 from maasservicelayer.context import Context
 
 logger = structlog.getLogger()
-
-UPDATE_OPERATION_STATUS_TIMEOUT = timedelta(seconds=30)
 
 
 def activity_defn_with_context(name):
@@ -64,78 +60,6 @@ def workflow_run_with_context(func):
         return res
 
     return wrapper
-
-
-def track_operation_status(func):
-    """Decorate a workflow run method to track its operation status in the DB."""
-    from maastemporalworker.workflow.operation import (
-        OPERATION_UUID_SEARCH_ATTRIBUTE,
-        UPDATE_OPERATION_STATUS_ACTIVITY_NAME,
-        UpdateOperationStatusParam,
-    )
-
-    @wraps(func)
-    async def wrapper(self, param):
-        info = workflow.info()
-        operation_uuid = info.search_attributes.get(
-            OPERATION_UUID_SEARCH_ATTRIBUTE
-        )
-        if not operation_uuid:
-            raise ApplicationError(
-                f"Status tracking is enabled for workflow {info.workflow_type}"
-                f" but the search attribute {OPERATION_UUID_SEARCH_ATTRIBUTE}"
-                " has not been set."
-            )
-        # Search attributes are always returned as a list.
-        operation_uuid = str(operation_uuid[0])
-
-        try:
-            await workflow.execute_local_activity(
-                UPDATE_OPERATION_STATUS_ACTIVITY_NAME,
-                UpdateOperationStatusParam(
-                    operation_uuid=operation_uuid,
-                    status=OperationStatus.RUNNING,
-                ),
-                start_to_close_timeout=UPDATE_OPERATION_STATUS_TIMEOUT,
-            )
-
-            result = await func(self, param)
-
-            await workflow.execute_local_activity(
-                UPDATE_OPERATION_STATUS_ACTIVITY_NAME,
-                UpdateOperationStatusParam(
-                    operation_uuid=operation_uuid,
-                    status=OperationStatus.COMPLETED,
-                ),
-                start_to_close_timeout=UPDATE_OPERATION_STATUS_TIMEOUT,
-            )
-            return result
-        except Exception as e:
-            logger.error(
-                "Workflow failed",
-                operation_uuid=operation_uuid,
-                error=str(e),
-            )
-            await workflow.execute_local_activity(
-                UPDATE_OPERATION_STATUS_ACTIVITY_NAME,
-                UpdateOperationStatusParam(
-                    operation_uuid=operation_uuid,
-                    status=OperationStatus.FAILED,
-                    error=str(e),
-                ),
-                start_to_close_timeout=UPDATE_OPERATION_STATUS_TIMEOUT,
-            )
-            raise
-
-    return wrapper
-
-
-def workflow_run_with_tracked_operation(func):
-    """
-    You MUST use this decorator instead of @workflow_run_with_context
-    on operation-tracked workflows.
-    """
-    return workflow_run_with_context(track_operation_status(func))
 
 
 def async_retry(retries=5, backoff_ms=1000):
